pyproject.toml
[project]
name = "azure-ai-foundary-project-python"
version = "0.1.0"
description = "Python template for Azure AI Foundry projects"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-projects>=1.0.0b12",
    "azure-identity>=1.23.0",
    "azure-search-documents>=11.5.3",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.3",
    "ty>=0.0.1a14",
]
